+++ src/console/program/OnlineCourse.java
@@ -2,7 +2,7 @@ package console.program;
 
 public class OnlineCourse extends Course {
 
-    public OnlineCourse(String courseName, String year, String mode, String dayOfLecture, int startTime, int duration) {
+    public OnlineCourse(String courseName, String year, String mode, String dayOfLecture, int startTime, double duration) {
         super(courseName, year, mode, dayOfLecture, startTime, duration);
 
     }
